+++ clang-tools-extra/docs/ReleaseNotes.rst
@@ -105,6 +105,11 @@ New checks
   Finds usages of ``realloc`` where the return value is assigned to the
   same expression as passed to the first argument.
 
+- New :doc:`bugprone-unused-ntd-object
+  <clang-tidy/checks/bugprone/unused-ntd-object>` check.
+
+  Finds unused variables with nontrivial destructors that are unlikely to be used as RAII.
+
 - New :doc:`cppcoreguidelines-avoid-const-or-ref-data-members
   <clang-tidy/checks/cppcoreguidelines/avoid-const-or-ref-data-members>` check.